Update index.ts
The Index.ts - the initial file responsible for starting the app by calling the SetupServer class - will start the server with the 'init' method. Since 'init' returns a 'promise' it is necessary to put 'await' on it, and 'await' has to be inside an async function. Since nobody is going to call this file, a Self-Invoking Functions is used.