Add settings for different deployment environments
The default.yml file is the default config file and refers to the development environment.
The test.yml file is the configuration for the test environment and in addition to having all the values of default.yml overrides the one that is defined.
The custom-environment-variables.yml map the environment variable names into your configuration structure, override all configuration files and Only command line options take precedence over them.