Add graceful shutdown
An app's processes can be started or stopped at any time, so the app must be prepared and know what to do when its services are interrupted. Using IPCs [Inter-Process Communication], it is possible to warn the app and define a behavior when an action occurs, for example to terminate processes and exit the program.
Upon receiving the signals ['SIGINT', 'SIGTERM', 'SIGQUIT'] the app will terminate connections to the database and to the server.