Deal with unhandled rejection and uncaught exception
The 'unhandledRejection' event is emitted whenever a Promise is rejected and no error handler is attached to the promise, note that promises can be captured and treated with catch(). The app will shut down throwing an exception to end the process.
About unhandled exceptions, the Node.js documentation says: 'Unhandled exceptions inherently mean that an application is in an undefined state. Attempting to resume application code without properly recovering from the exception can cause additional unforeseen and unpredictable issues.'
When an unhandled exception is thrown in the app, the 'uncaughtException' event is triggered, then the app logs the exception, ends the failed process and then shuts down.