Add globals.d.ts
Compose a type that already exists with another type, add a testRequest type to the global.
Note: if you add the import {Supertest} from 'supertest' and type testRequest: SuperTest <import ('supertest'). Test>, the typescript will treat this file as a local and not a global module. To be global and to allow overriding global types, it must be declared inline.