REFACTOR Should successfully create a new orphan hosting
> Create the OrphanHosting model by defining the interface and the schema.
The toJSON method is standard of mongoose, it converts the mongoose document into a simple JavaScript object, so if an object has a toJSON function, JSON.stringify calls toJSON and serializes the return value of toJSON, that is, serializes the model that is fetched from the database. And the transform method defines how the data will be transformed when moving from a mongoose document to a JSON document.
> Update test to clean the orphanhostings collection from the database before each test and add objectContaining in expect so that the expected object is a subset of the received object, and not identical objects.
> Update the OrphanHostingController controller to save an object in the database.