Add multer and your types
The form to be presented to the user for insertion of orphan hostings will contain information, either generated by the user's input or included from files that the user selected, which will be sent to the server using the POST method. The HTML forms provide the enctype attribute that specifies how the data should be encoded when sending it to the server, the multipart / form-data was used because we will upload files.
Multer is a package that handles multipart/form-data, he adds a body object and a file or files object to the request object, where the body object contains the values of the text fields of the form, the file or files object contains the files uploaded via the form.