Add ArchetypeError.ts
This file will contain an interface called 'archetypeError' where you will have the code and message, and optionally description and documentation. The 'name' property, corresponding name of the status code and obtained using the http-status module, will be on another interface called 'archetypeErrorResponse' which will extend the properties of the first.
You will also have a class called 'ArchetypeError' with a method that will receive a structure of type archetypeError and fill the properties by returning a structure of type archetypeErrorResponse.