Add jsonwebtoken and your type
jsonwebtoken is a JSON Web Token [JWT] library, which is a standard used to create access tokens for an application. The user signs in using his email and password, the server verifies the user's identity and sends a signed jwt to the user. Then, the user uses the token to access protected resources, passing the JWT in the HTTP authorization header, so the server knows that the user is authenticated.