RED should return UNAUTHORIZED if there is a problem on the token verification
The test seeks to throw the unauthorized error when there is an invalid token. We will run the authentication middleware with its mocked arguments.
-> reqFake that represents the request contains in the headers authorization the invalid token;
-> resFake that represents the response contains a mock version of status and send;
-> nextFake is a mock version of next().
When executing the authentication middleware function, your request will contain the invalid token, the error will occur when trying to decode the token and we must send a response with status 401 and the error formatted in the response body.
The current test run returns the error: 'JsonWebTokenError: jwt malformed'