RED should return CONFLICT when have the email already registered
A user is saved, I create another user with the same email and send a request post('/user/create'), when trying to save, a search is made if there is already a user with the same email and if it does, it will launch a validation error duplicate email address. The controller will catch this error and format it using sendValidationErrorResponse of NestErrors.
The test fails giving error MongoError: duplicate key and will have 500 status because the 409 error has not yet been treated.