Handle MongoError for duplicate key
According to the mongoose documentation the unique option for schemas is not a validator but a helper for building MongoDB unique indexes, that is, if MongoDB does not yet have a unique index in a field it will still be possible to save it in the database even if it has unique as true.
Using SchemaType.prototype.validate() creates a validator function that searches the database if it already has the same email and that returns false indicating that the validation failed, creates the message function that returns the error message as a string and also defines a type to differentiate from other mongoose validation errors. So the test will have ValidationError instead of MongoError.