RED should check if the user has an administrator role and call the next middleware
The test seeks to test the behavior of middleware functions in relation to the task of calling the next middleware on the stack in the authentication middleware. We saved a user with a role: ['adm'], created a fake request containing the user's id, mocked next function and then called the middleware 'adm', after executing 'adm' we checked if the mocked next function was called.