RED should return FORBIDDEN if there is no role adm
The test seeks to throw the forbidden error when there is no role adm in user. We will run the adm middleware with its mocked arguments.
-> reqFake that represents the request contains in the decoded.id the id of user;
-> resFake that represents the response contains a mock version of status and send;
-> nextFake is a mock version of next().
When executing the middleware function contain the params above, an error will occur when checking roles and we must send a response with status 403 and the error formatted in the response body.