RED should generate a token for a valid user to reset their password
Mock the new Date() constructor to define a date determined by me, use jest.spyOn in Date so that when it is invoked it uses the mocked version. I use mockRestore at the end of the operation to resume normal operation.
When saving a user in the database, Date's 'now' method is used, but as Date was mocked the test returned: 'ValidationError: user validation failed: _id: Date.now is not a function'. So mock also Date's now method, resuming its normal operation at the end of the operation.
When sending POST request to '/user/forgot-password' with authorization and the email field, the controller must return a token containing your email and expiration date.