REFACTOR should generate a token for a valid user to reset their password
Import the email instance created in 'src/modules/mailer.ts', and use the 'send' method to send an email template, the template property indicates the folder where the email body will be sent, inside the auth folder has the html.pug file which has interpolations in its content to add the custom username and link indicated in the configuration as locals.
This link will be of the type `http://localhost:3000/reset-password?t=${token}`, which will be used to reset the password containing the token in query params.