REFACTOR handle http and custom errros
Separate http errors and custom errors to facilitate error handling.
Each HTTP error will inherit from its error class (1xx, 2xx, 3xx, 4xx, 5xx) which inherits from the HttpError class which inherits from Error. The HttpError class contains the properties of the http error and a method that sends a response with the status code and the error formatted by ArchetypeError.
Each custom error will inherit from CustomErros that inherits from Error. The CustomErros class has a method that sends a response with the error message.
Using instanceof it is possible to identify whether the error is of the type HttpError or CustomError.